实现mybatis持久层通用Mapper,  搭建web环境

+++ core/src/main/java/com/ace/core/persistence/mapper/GenericeMapper.java
@@ -26,7 +26,7 @@ public interface GenericeMapper<T, ID extends Serializable>  {
      * @return
      *              返回保存实体
      */
-    public Long insert(T entity);
+    public int insert(T entity);
 
 
     /**
